refactor: 통신 기능 추가 형태로 변경
