+++ NowPlayingList/NowPlayingList/Extensions/JSONDecoder+.swift
@@ -1,0 +1,15 @@
+//
+//  JSONDecoder+.swift
+//  NowPlayingList
+//
+//  Created by Yongwoo Marco on 2022/02/26.
+//
+
+import Foundation
+
+extension JSONDecoder {
+    convenience init(keyDecodingStrategy: JSONDecoder.KeyDecodingStrategy) {
+        self.init()
+        self.keyDecodingStrategy = keyDecodingStrategy
+    }
+}
